styles and restructure
