+++ package.json
@@ -18,11 +18,13 @@
   "devDependencies": {
     "@types/jest": "^29.5.3",
     "@types/p5": "^1.6.2",
+    "css-loader": "^6.8.1",
     "gh-pages": "^5.0.0",
     "html-webpack-plugin": "^5.5.3",
     "jest": "^29.6.2",
     "jest-environment-jsdom": "^29.6.2",
     "p5": "^1.7.0",
+    "style-loader": "^3.3.3",
     "ts-jest": "^29.1.1",
     "ts-loader": "^9.4.4",
     "typescript": "^5.1.6",
