+++ src/index.html
@@ -1,0 +1,57 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/4.7.0/css/font-awesome.min.css">
+    <meta charset="UTF-8">
+    <title>Ant Colony Simulation</title>
+</head>
+
+<body>
+    <div class="control-panel-button-container">
+        <a href="javascript:void(0);" class="control-panel-button icon" onclick="openNav()">
+            <i class="fa fa-bars"></i>
+        </a>
+    </div>
+
+    <div class="control-panel" class="sidenav">
+        <div class="control-panel-content">
+            <a href="javascript:void(0)" class="close-button" onclick="closeNav()">&times;</a>
+            <div class="control-panel-slider-container">
+                <span class="control-panel-slider-label">maxSpeed</span>
+                <input class="control-panel-slider" id="maxSpeed" type="range" min="0.1" max="5.0" value="2.5" step="0.1">
+            </div>
+            <!-- <a href="#">Clients</a> -->
+
+        </div>
+    </div>
+
+</body>
+<script>
+    function openNav() {
+        document.getElementsByClassName("control-panel")[0].style.width = "250px";
+        document.getElementsByClassName("control-panel-button")[0].style.opacity = "0";
+
+    }
+
+    function closeNav() {
+        document.getElementsByClassName("control-panel")[0].style.width = "0px";
+        document.getElementsByClassName("control-panel-button")[0].style.opacity = "1";
+    }
+    // const myDiv = document.getElementById('nav');
+    // const myCanvas = document.getElementById('defaultCanvas0');
+
+    // myDiv.addEventListener('mouseenter', () => {
+    //     const myCanvas = document.getElementById('defaultCanvas0');
+    //     myCanvas.style.pointerEvents = 'none';
+    // });
+
+    // myDiv.addEventListener('mouseleave', () => {
+    //     const myCanvas = document.getElementById('defaultCanvas0');
+    //     // myCanvas.style.pointerEvents = 'auto';
+    //     myCanvas.style.pointerEvents = 'none';
+    //     myCanvas.style.cursor = "not-allowed";
+    // });
+</script>
+
+</html>
